add uninstall script

+++ scripts/restart.sh
@@ -5,13 +5,14 @@
 if ! grep -q "# NIDS" /etc/hosts; then
     echo -e "\n# NIDS\n127.0.0.1 mongo" | sudo tee -a /etc/hosts
     echo "NIDS entry added to /etc/hosts"
-else
-    echo "NIDS entry already exists in /etc/hosts"
 fi
 
-echo "Cleaning Docker Environment..."
-docker-compose --profile feeder --profile gui down
-docker system prune --volumes -af
+read -p Do you want to clean the Docker environment? \(y/n\): clean_docker
+if [[ "$clean_docker" == "y" ]]; then
+    echo "Cleaning Docker Environment..."
+    docker-compose --profile feeder --profile gui down
+    docker system prune --volumes -af
+fi
 
 read -p "Do you want to clear mongo-db data? (y/n): " clear_mongo
 if [[ "$clear_mongo" == "y" ]]; then
@@ -20,7 +21,5 @@ if [[ "$clear_mongo" == "y" ]]; then
 fi
 
 echo "-------------------------------------------------------------------"
-echo "Starting Services..."
+echo "Restarting Services..."
 docker-compose --profile feeder --profile gui up
-
-echo "Restarted."
